chore: clean up project structure and fix peripheral issues
- Add docs/DEV.md with development plan and architecture comparison
- Remove xcompiler symlink, empty config/isa32.rs, isa64.rs, device/mod.rs
- Remove unused atomic.rs and zifence.rs instruction stubs
- Remove unused bitflags dependency from xcore
- Fix xlogger xprintln!/xprint! macros to actually print (were returning
  format_args! without side effects)
- Remove unused Write impl for Logger, simplify log coloring
- Simplify xdb error handling and CLI dispatch

+++ xemu/xdb/src/cli.rs
@@ -18,7 +18,7 @@ enum Commands {
     Step {
         /// Number of instructions to step
         #[arg(default_value_t = 1)]
-        count: u32,
+        count: u64,
     },
     /// Continue execution
     #[command(alias = "c")]
